Code to automatically set fire stick resolution
The fire stick 4K occasionally resets its resolution to default. This
code will check the resolution and set it to the correct one, which must
be specified in the conf file for each fire stick.

The check is done every time leancap_ready starts and once per day while
it is running.

This relies on the fire stick settings application and its peculiar user
interface. If that changes or the home page changes it will no longer work.